import everything from colink & small fixes
