+++ colink/variable_transfer.py
@@ -6,6 +6,7 @@ import colink as CL
 def set_variable(self, key: str, payload: bytes, receivers: List[CL.Participant]):
     if self.task_id is None:
         logging.error("set_variable task_id not found")
+        raise Exception("set_variable task_id not found")
     new_participants = [CL.Participant(user_id=self.get_user_id(), role="requester")]
     for p in receivers:
         if p.user_id == self.get_user_id():
@@ -33,6 +34,7 @@ def set_variable(self, key: str, payload: bytes, receivers: List[CL.Participant]
 def get_variable(self, key: str, sender: CL.Participant) -> bytes:
     if self.task_id is None:
         logging.error("get_variable task_id not found")
+        raise Exception("get_variable task_id not found")
     key = "_remote_storage:private:{}:_variable_transfer:{}:{}".format(
         sender.user_id, self.get_task_id(), key
     )
