+++ examples/host_import_users_and_set_registry.py
@@ -1,12 +1,7 @@
 import sys
 import logging
 import colink as CL
-from colink import CoLink
-from colink.sdk_a import (
-    get_time_stamp,
-    generate_user,
-    prepare_import_user_signature,
-)
+from colink import CoLink, get_time_stamp, generate_user, prepare_import_user_signature
 
 if __name__ == "__main__":
     logging.basicConfig(
