+++ examples/user_lock.py
@@ -1,8 +1,7 @@
 import sys
 import logging
 import threading
-from colink import CoLink
-from colink.sdk_a import byte_to_int
+from colink import CoLink, byte_to_int
 
 
 def lock_and_unlock(cl):
