+++ examples/user_wait_task.py
@@ -1,7 +1,6 @@
 import sys
 import colink as CL
-from colink import CoLink
-from colink.sdk_a import decode_jwt_without_validation, str_to_byte
+from colink import CoLink, decode_jwt_without_validation
 
 if __name__ == "__main__":
     addr = sys.argv[1]
